code motion
